requirements.txt
cloudinary==1.39.1
Django==5.0
django_crispy_forms==2.1
django_resized==1.0.2
dj_database_url
geopy
python-dotenv
gunicorn
Pillow==10.2.0
crispy_bootstrap5
django-filter
psycopg
whitenoise